format

+++ LydighedAPI/Controllers/CourseController.cs
@@ -55,7 +55,7 @@ namespace API.Controllers
         [HttpPut]
         public async Task<IActionResult> UpdateCourse([FromBody] UpdateCourseViewModel updateCourseViewModel)
         {
-            if(updateCourseViewModel is null) return BadRequest("ViewModel was null");
+            if (updateCourseViewModel is null) return BadRequest("ViewModel was null");
 
             Course updateCourse = new Course(updateCourseViewModel.UpdatedCourseId, updateCourseViewModel.Level);
             await _courseRepository.UpdateCourse(updateCourse);
@@ -65,9 +65,13 @@ namespace API.Controllers
 
 
         [HttpDelete]
-        public async Task DeleteCourse(int courseId)
+        public async Task<IActionResult> DeleteCourse(int courseId)
         {
+            if (courseId <= 0) return BadRequest("CourseId must be larger than zero");
+
             await _courseRepository.DeleteCourse(courseId);
+            
+            return Ok();
         }
     }
 }
